[fix]すべての"creation"を"work"にすべての"Creation"を"Work"に変更
