+++ app/src/main/java/com/example/funcy_portfolio_android/ui/workRegister/WorkRegisterFragment.kt
@@ -1,4 +1,4 @@
-package com.example.funcy_portfolio_android.ui.creationRegister
+package com.example.funcy_portfolio_android.ui.workRegister
 
 import android.app.AlertDialog
 import android.content.Context
@@ -19,14 +19,12 @@ import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.funcy_portfolio_android.R
-import com.example.funcy_portfolio_android.databinding.FragmentCreationRegisterBinding
+import com.example.funcy_portfolio_android.databinding.FragmentWorkRegisterBinding
 import com.example.funcy_portfolio_android.databinding.ItemTagBinding
-import com.google.android.material.chip.Chip
-import com.google.android.material.chip.ChipGroup
 
-class CreationRegisterFragment : Fragment() {
-    private val viewModel: CreationRegisterViewModel by activityViewModels()
-    private lateinit var binding: FragmentCreationRegisterBinding
+class WorkRegisterFragment : Fragment() {
+    private val viewModel: WorkRegisterViewModel by activityViewModels()
+    private lateinit var binding: FragmentWorkRegisterBinding
 
     //画像選択
     private var image_launchar = registerForActivityResult(ActivityResultContracts.OpenMultipleDocuments()) {
@@ -37,7 +35,7 @@ class CreationRegisterFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentCreationRegisterBinding.inflate(inflater, container, false)
+        binding = FragmentWorkRegisterBinding.inflate(inflater, container, false)
 
         val tags = viewModel.resetTag()
         Log.e("resetTag",tags.toString())
@@ -45,7 +43,7 @@ class CreationRegisterFragment : Fragment() {
         //画像変更
         viewModel.thumbnail.observe(viewLifecycleOwner, Observer {
             binding.rlThumbnail.layoutManager = LinearLayoutManager(context,LinearLayoutManager.HORIZONTAL,false)
-            binding.rlThumbnail.adapter = viewModel.thumbnail.value?.let { CreationRegisterCardAdapter(it) }
+            binding.rlThumbnail.adapter = viewModel.thumbnail.value?.let { WorkRegisterCardAdapter(it) }
         })
 
         return binding.root
@@ -55,7 +53,7 @@ class CreationRegisterFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding.btBack.setOnClickListener {
-            findNavController().navigate(R.id.action_CreationRegisterFragment_to_MainFragment)
+            findNavController().navigate(R.id.action_WorkRegisterFragment_to_MainFragment)
         }
 
         binding.swPublic.setOnCheckedChangeListener { buttonView, isChecked ->
@@ -66,11 +64,11 @@ class CreationRegisterFragment : Fragment() {
         }
 
         binding.btAddTag.setOnClickListener {
-            val creationRegisterBottomSheet = CreationRegisterBottomSheet()
+            val workRegisterBottomSheet = WorkRegisterBottomSheet()
             activity?.let {
-                creationRegisterBottomSheet.show(
+                workRegisterBottomSheet.show(
                     it.supportFragmentManager,
-                    CreationRegisterBottomSheet.TAG
+                    WorkRegisterBottomSheet.TAG
                 )
             }
         }
@@ -95,19 +93,19 @@ class CreationRegisterFragment : Fragment() {
 
         /* 作品投稿処理 */
         binding.btRegister.setOnClickListener{
-            if(binding.etCreationTitle.text.toString() != "") {
+            if(binding.etWorkTitle.text.toString() != "") {
                 AlertDialog.Builder(activity)
                     .setTitle("作品を投稿しますか？")
                     .setPositiveButton("登録する", DialogInterface.OnClickListener { dialog, which ->
-                        viewModel.registerCreation(
-                            binding.etCreationTitle.text.toString(),
-                            binding.etCreationDescription.text.toString(),
+                        viewModel.registerWork(
+                            binding.etWorkTitle.text.toString(),
+                            binding.etWorkDescription.text.toString(),
                             1,
                             binding.etGitHubLink.text.toString(),
                             binding.etYoutubeLink.text.toString()
                         )
 
-                        findNavController().navigate(R.id.action_CreationRegisterFragment_to_MainFragment)
+                        findNavController().navigate(R.id.action_WorkRegisterFragment_to_MainFragment)
 
                     })
                     .setNegativeButton("キャンセル", null)
@@ -115,7 +113,7 @@ class CreationRegisterFragment : Fragment() {
 
             }else{
                 binding.otfTitle.isErrorEnabled = true
-                binding.otfTitle.error = getString(R.string.creation_error_title_null)
+                binding.otfTitle.error = getString(R.string.work_error_title_null)
             }
         }
     
@@ -124,10 +122,10 @@ class CreationRegisterFragment : Fragment() {
         }
 
         //フォーカスが外れたとき→別の場所おしたときキーボード隠す
-        binding.etCreationTitle.setOnFocusChangeListener { _, hasFocus ->
+        binding.etWorkTitle.setOnFocusChangeListener { _, hasFocus ->
             if(!hasFocus)showoffKeyboard()
         }
-        binding.etCreationDescription.setOnFocusChangeListener { _, hasFocus ->
+        binding.etWorkDescription.setOnFocusChangeListener { _, hasFocus ->
             if(!hasFocus)showoffKeyboard()
         }
         binding.etGitHubLink.setOnFocusChangeListener { _, hasFocus ->
@@ -138,7 +136,7 @@ class CreationRegisterFragment : Fragment() {
         }
 
         //Enterキー押したときキーボード隠す
-        binding.etCreationTitle.setOnEditorActionListener { _, i, keyEvent ->
+        binding.etWorkTitle.setOnEditorActionListener { _, i, keyEvent ->
             return@setOnEditorActionListener getActions(i,keyEvent)
         }
         binding.etGitHubLink.setOnEditorActionListener { _, i, keyEvent ->
