+++ app/src/main/res/drawable/img_work_detail_thumbnail.xml
