+++ app/src/main/res/layout/fragment_group_mypage.xml
@@ -90,7 +90,7 @@
             <com.google.android.material.tabs.TabItem
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/creation_list" />
+                android:text="@string/work_list" />
 
             <com.google.android.material.tabs.TabItem
                 android:layout_width="wrap_content"
@@ -128,7 +128,7 @@
             app:layout_constraintTop_toBottomOf="@+id/tabLayout">
 
             <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/rvGroupCreationList"
+                android:id="@+id/rvGroupWorkList"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content" />
         </androidx.core.widget.NestedScrollView>
