+++ app/src/main/res/layout/fragment_group_register.xml
@@ -91,7 +91,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="8dp"
-                android:text="@string/creation_necessary"
+                android:text="@string/work_necessary"
                 android:textColor="@color/red"
                 app:layout_constraintStart_toEndOf="@+id/tvGroupName"
                 app:layout_constraintTop_toTopOf="@+id/tvGroupName" />
@@ -128,7 +128,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="8dp"
-                android:text="@string/creation_necessary"
+                android:text="@string/work_necessary"
                 android:textColor="@color/red"
                 app:layout_constraintStart_toEndOf="@+id/tvGroupLeaderAddress"
                 app:layout_constraintTop_toTopOf="@+id/tvGroupLeaderAddress" />
@@ -187,7 +187,7 @@
                 android:background="@null"/>
 
             <Button
-                android:id="@+id/btRegisterGroupCreation"
+                android:id="@+id/btRegisterGroupWork"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="直で作品を投稿"
@@ -202,7 +202,7 @@
                 android:text="グループ作品一覧へ"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/btRegisterGroupCreation" />
+                app:layout_constraintTop_toBottomOf="@+id/btRegisterGroupWork" />
 
 
         </androidx.constraintlayout.widget.ConstraintLayout>
